Create _212-Sort an array of 0s, 1s and 2s
(For Best Explanation :https://youtu.be/Kov-M_lTuOA)

Leetcode(Sort Colours):
Given an array nums with n objects colored red, white, or blue, sort them in-place so that objects of the same color are adjacent, with the colors in the order red, white, and blue.
We will use the integers 0, 1, and 2 to represent the color red, white, and blue, respectively.
You must solve this problem without using the library's sort function.
                                    
                                     OR
 
 GFG(Sort an array of 0s, 1s and 2s):                                                                OR
Given an array A[] consisting 0s, 1s and 2s. The task is to write a function that sorts the given array. The functions should put all 0s first, then all 1s and all 2s in last.