Update _234-Set the K-th bit of a given number
Given a number n and a value k. From the right, Set(/Turn on) the Kth bit in the binary representation of n. Please note that k = 1 means the rightmost bit. Also, 1 <= k <= x, where x is the number of bits in the binary representation of n.
Examples: 
 

Input : n = 10, k = 2
Output : 10
(2nd bit of 10 is already set.)


Input : n = 17, k = 3
Output : 21