Update _222-Find Duplicates in an array in O(n) and by using O(1) extra space
Given an Integer array of n elements where all the Elements of array are in the range Of 1 to n, with any of these numbers appearing any number of times. Find these repeating numbers in O(n) and using only constant memory space.(O(1) Extra Space)

Example:
Input : n = 7 and array[] = {1, 7, 3, 7, 3, 7, 1}
Output: 1, 3, 7

Explanation: The numbers 1 , 3 and 7 appears more
than once in the array.