Update and rename _248-Position of rightmost different bit to _235-Position Of Rightmost Different Bit
Given two numbers m and n. Find the position of the rightmost different bit in the binary representation of numbers. It is guaranteed that such a bit exists. 

Examples: 

Input: m = 11, n = 9
Output: 2
(11)10 = (1011)2
(9)10 = (1001)2
It can be seen that 2nd bit from
the right is different 

Input: m = 52, n = 4
Output: 5
(52)10 = (110100)2
(4)10 = (100)2, can also be written as
     = (000100)2
It can be seen that 5th bit from
the right is different