Update _308-Find A Triplet That Sum To A Given Value
Given an array and a value, find if there is a triplet in array whose sum is equal to the given value. If there is such a triplet present in array then print the triplet ,If there is no triplet, then print that no triplet exist.

Examples: 
 
Input: array = {12, 3, 4, 1, 6, 9}, sum = 24; 
Output: (12, 3, 9) 
Explanation: There is a triplet (12, 3 and 9) present in the array whose sum is 24. 

Input: array = {1, 2, 3, 4, 5}, sum = 9 
Output: (5, 3, 1)
Explanation: There is a triplet (5, 3 and 1) present in the array whose sum is 9.