Update _324-Replace Every Array Element By Bitwise XOR Of Previous And next Element
Given an array of integers, replace every element with xor of previous and next elements with following exceptions. 
a) First element is replaced by sum of first and second. 
b) Last element is replaced by sum of last and second last.

Examples: 

Input: arr[] = { 2, 3, 4, 5, 6}
Output: 1 6 6 2 3 

We get the following array as {2^3, 2^4, 3^5, 4^6, 5^6}

Input: arr[] = { 1, 2, 1, 5}
Output: 3, 0, 7, 4

We get the following array as {1^2, 1^1, 2^5, 1^5}