Rename _191-Sort an array of 0s, 1s and 2s. to _191- Sort an array of 0s, 1s and 2s.
(For Best Explanation :https://youtu.be/Kov-M_lTuOA)

Leetcode 75 : Sort Colors(Sort Colours):
Given an array nums with n objects colored red, white, or blue, sort them in-place so that objects of the same color are adjacent, with the colors in the order red, white, and blue.
We will use the integers 0, 1, and 2 to represent the color red, white, and blue, respectively.
You must solve this problem without using the library's sort function.

Input :    [1,1,0,0,2,0,1,2,1,0]
Output : [0,0,0,01,1,1,1,2,2]

Input :    [2,0,1] //Reason for i<=p2
Output : [0,1,2]
                                                                                           
                                                                                                                     OR
 GFG(Sort an array of 0s, 1s and 2s):                                                              
Given an array A[] consisting 0s, 1s and 2s. The task is to write a function that sorts the given array. The functions should put all 0s first, then all 1s and all 2s